ctrd: try to unpack image when create snapshot
PouchContainer always unpacks for pulled images. But if there is any
crash or terminated, the unpack action will be stopped. And image
metadata is stored in containerd. It will cause that the following
create container request fails. Try to unpack it if fail.

Signed-off-by: Wei Fu <fuweid89@gmail.com>


Rebased from commit 5718d0d: ctrd: try to unpack image when create snapshot